refactoring forms and menu
